pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "campus-events"
version = "0.1.0"
description = "Automation-safe campus events digest pipeline"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0",
]

[tool.setuptools.packages.find]
include = ["campus_events*"]

